LayoutTests/media/session/transient-solo-interruptions-expected.txt

'Transient Solo' interruption events should pause elements belonging to 'Content' media sessions.


Waiting for Content media to begin playing.
EVENT(canplaythrough)
Media began playing.
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
EXPECTED (video.paused == 'false') OK
Sending 'Transient Solo' start-of-interruption notification.
RUN(internals.sendMediaSessionStartOfInterruptionNotification("transient-solo"))
EXPECTED (internals.mediaSessionCurrentState(session) == 'interrupted') OK
EXPECTED (video.paused == 'true') OK
Sending 'Transient Solo' end-of-interruption notification.
RUN(internals.sendMediaSessionEndOfInterruptionNotification("transient-solo"))
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
EXPECTED (video.paused == 'false') OK
END OF TEST
